+++ classes/Application.php
@@ -27,14 +27,14 @@ use Psr\Cache\CacheItemPoolInterface;
 class Application extends Hookable implements Interface_Theme, Interface_Factory {
 	
 	/**
-	 * Equivalent of zesk()
+	 * Probably should discourage use of this. Zesk singleton.
 	 *
 	 * @var Kernel
 	 */
 	public $zesk = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit the value here.
 	 *
 	 * @var Paths
@@ -42,7 +42,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $paths = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Hooks
@@ -50,7 +50,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $hooks = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Autoloader
@@ -58,7 +58,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $autoloader = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var CacheItemPoolInterface
@@ -66,7 +66,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $cache = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Configuration
@@ -79,7 +79,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 */
 	public $loader = null;
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Logger
@@ -87,7 +87,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $logger = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Classes
@@ -95,7 +95,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $classes = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Objects
@@ -103,7 +103,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	public $objects = null;
 	
 	/**
-	 * Inherited directly from zesk().
+	 * Inherited directly from zesk\Kernel.
 	 * Do not edit.
 	 *
 	 * @var Process
@@ -800,7 +800,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return unknown|object|\zesk\NULL|mixed
 	 */
 	final public function object_singleton($class) {
-		zesk()->deprecated();
+		$this->deprecated();
 		$args = func_get_args();
 		$args[0] = $this;
 		$object = $this->call_hook_arguments("singleton_$class", $args, null);
@@ -1877,7 +1877,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return string
 	 */
 	public function url($uri) {
-		zesk()->deprecated();
+		$this->deprecated();
 		// TODO Remove this
 		return $uri;
 	}
@@ -1939,21 +1939,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 			"method" => $name
 		));
 	}
-	/**
-	 * Load the Application singleton
-	 *
-	 * @deprecated 2017-08 Globals are bad.
-	 * @param array $options
-	 * @throws Exception_Configuration
-	 * @return Application
-	 * @todo this should be called "singleton" but that call is used for creating singleton objects
-	 *       in the application. So deprecate that first, then deprecate this once that's gone.
-	 */
-	public static function instance(array $options = array()) {
-		zesk()->deprecated();
-		return zesk()->application();
-	}
-	
+
 	/**
 	 * @deprecated 2017-12
 	 * @param mixed $add
@@ -2026,7 +2012,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Ambigous <mixed, array>
 	 */
 	public function _class_cache($class, $component = "") {
-		zesk()->obsolete();
+		$this->zesk->obsolete();
 	}
 	
 	/**
@@ -2036,7 +2022,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return ORM
 	 */
 	public final function object($class, $mixed = null, $options = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->orm_registry($class, $mixed, $options);
 	}
 	
@@ -2050,7 +2036,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return string|\zesk\Ambigous
 	 */
 	public final function object_table_name($class, $mixed = null, $options = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class, $mixed, $options)->table();
 	}
 	
@@ -2064,7 +2050,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return string|\zesk\Ambigous
 	 */
 	public final function object_table_columns($class, $mixed = null, $options = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class, $mixed, $options)->columns();
 	}
 	
@@ -2078,7 +2064,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return \zesk\Database
 	 */
 	public final function object_database($class, $mixed = null, $options = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class, $mixed, $options)->database();
 	}
 	
@@ -2088,7 +2074,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Database_Query_Select
 	 */
 	public function query_select($class, $alias = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class)->query_select($alias);
 	}
 	/**
@@ -2096,7 +2082,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Database_Query_Update
 	 */
 	public function query_update($class, $alias = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class)->query_update($alias);
 	}
 	/**
@@ -2105,7 +2091,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Database_Query_Insert
 	 */
 	public function query_insert($class) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class)->query_insert();
 	}
 	/**
@@ -2113,7 +2099,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Database_Query_Insert
 	 */
 	public function query_insert_select($class, $alias = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class)->query_insert_select($alias);
 	}
 	
@@ -2122,7 +2108,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Database_Query_Delete
 	 */
 	public function query_delete($class) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->object($class)->query_delete();
 	}
 	
@@ -2133,7 +2119,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return Class_ORM
 	 */
 	public function class_orm($class) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->class_orm_registry($class);
 	}
 	
@@ -2144,7 +2130,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return \zesk\Database
 	 */
 	public final function class_orm_database($class) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->orm_registry($class)->database();
 	}
 	
@@ -2154,7 +2140,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @throws Exception_Parameter
 	 */
 	public function clear_class_cache($class = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		
 		return $this->orm_registry()->clear_cache($class);
 	}
@@ -2165,7 +2151,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @param unknown $add
 	 */
 	final public function orm_classes($add = null) {
-		zesk()->deprecated();
+		$this->deprecated();
 		
 		return $this->modules->object("orm")->orm_classes($add);
 	}
@@ -2180,7 +2166,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return array
 	 */
 	final public function all_classes() {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->modules->object("orm")->all_classes();
 	}
 	
@@ -2191,7 +2177,7 @@ class Application extends Hookable implements Interface_Theme, Interface_Factory
 	 * @return multitype:
 	 */
 	public function schema_synchronize(Database $db = null, array $classes = null, array $options = array()) {
-		zesk()->deprecated();
+		$this->deprecated();
 		return $this->modules->object("orm")->schema_synchronize($db, $classes, $options);
 	}
 }