+++ modules/orm/classes/ORM.php
@@ -199,7 +199,7 @@ class ORM extends Model {
 	 */
 	public function object_factory($class, $mixed = null, array $options = array()) {
 		$this->application->deprecated();
-		return self::orm_factory($class, $mixed, $options);
+		return $this->orm_factory($class, $mixed, $options);
 	}
 	
 	/**
@@ -211,8 +211,9 @@ class ORM extends Model {
 	 * @return \zesk\ORM
 	 */
 	public function orm_factory($class, $mixed = null, array $options = array()) {
-		return self::model_factory($class, $mixed, $options);
+		return $this->model_factory($class, $mixed, $options);
 	}
+	
 	/**
 	 * Create a new object
 	 *
@@ -240,29 +241,10 @@ class ORM extends Model {
 	}
 	
 	/**
-	 * Not sure why we're doing this; perhaps to force cyclical structures from being destroyed,
-	 * clean up memory references? KMD
-	 *
-	 * KMD Removed 2017-06-07, see:
-	 * 
-	 * https://stackoverflow.com/questions/2251113/should-i-use-unset-in-php-destruct
-	 * 
-	 * Monitor memory usage, how does PHP deal with cyclical references
-	 */
-	// 	public function __destruct() {
-	// 		foreach ($this->members as $k => $member) {
-	// 			unset($this->members[$k]);
-	// 		}
-	// 		$this->members = array();
-	// 		$this->class = null;
-	// 		$this->original = array();
-	// 	}
-	
-	/**
 	 * Wakeup functionality
 	 */
 	public function __wakeup() {
-		$this->application = zesk()->application();
+		$this->application = Kernel::singleton()->application();
 		$this->initialize_specification();
 		$this->initialize($this->members, 'raw');
 	}
@@ -2867,7 +2849,7 @@ class ORM extends Model {
 	 * @return Database_Query_Select
 	 */
 	function query($alias = null) {
-		zesk()->deprecated();
+		$this->application->deprecated();
 		return $this->query_select($alias);
 	}